Modulaize the Operations
