+++ src/config/api.js
@@ -1,0 +1,3 @@
+export const baseUrl = 'https://us-central1-js-capstone-backend.cloudfunctions.net/api/';
+
+export const ApiKey = 'qhvik7RAK2bFYB25gEBh';
