--- src/data.js
@@ -1,8 +1,0 @@
-export default [
-  { name: 'user1', score: 100 },
-  { name: 'user2', score: 200 },
-  { name: 'user3', score: 20 },
-  { name: 'user4', score: 100 },
-  { name: 'user5', score: 30 },
-  { name: 'user6', score: 600 },
-];
