+++ src/index.js
@@ -1,3 +1,5 @@
+import addScore from './module/addScore.js';
+import renderData from './module/renderData.js';
 import './style.css';
 
 let user = '';
@@ -22,27 +24,6 @@ const resetInput = () => {
   scoreField.value = '';
 };
 
-const baseUrl = 'https://us-central1-js-capstone-backend.cloudfunctions.net/api/';
-
-const ApiKey = 'qhvik7RAK2bFYB25gEBh';
-
-const scoreBoard = document.querySelector('.score-table');
-
-const renderData = async () => {
-  const response = await fetch(`${baseUrl}games/${ApiKey}/scores/`);
-  const { result } = await response.json();
-  const ul = document.getElementsByTagName('ul')[0];
-  ul.innerHTML = '';
-  result
-    .sort((a, b) => a.score - b.score)
-    .reverse()
-    .forEach((record) => {
-      const li = document.createElement('li');
-      li.innerHTML = `${record.user}:${record.score}`;
-      scoreBoard.appendChild(li);
-    });
-};
-
 document.addEventListener('DOMContentLoaded', () => renderData());
 
 const refreshBtn = document.querySelector('.refresh-btn');
@@ -53,18 +34,10 @@ const addBtn = document.querySelector('.add-btn');
 
 addBtn.addEventListener('click', async (e) => {
   e.preventDefault();
-
-  const response = await fetch(`${baseUrl}games/${ApiKey}/scores/`, {
-    method: 'POST',
-    body: JSON.stringify({ user, score: Number(score) }),
-    headers: {
-      'Content-type': 'application/json; charset=UTF-8',
-    },
+  addScore(user, score).then((result) => {
+    if (result === 'Leaderboard score created correctly.') {
+      renderData();
+      resetInput();
+    }
   });
-
-  const { result } = await response.json();
-  if (result === 'Leaderboard score created correctly.') {
-    renderData();
-    resetInput();
-  }
 });
